Ticket 46918 - Fix compiler warnings on arm
Description:  On "arm" architecture there were many compiler warnings
              and after fixing some it also adressed a crash in replication
              when trying to update the agreement maxcsn.

              All updatream architectures build without compilar warnings,
              and coverity scan did not report any regressions.

https://pagure.io/389-ds-base/issue/49618

Reviewed by: simon(Thanks!)

+++ ldap/servers/slapd/back-ldbm/filterindex.c
@@ -974,7 +974,7 @@ keys2idl(
     IDList *idl = NULL;
 
     slapi_log_err(SLAPI_LOG_TRACE, "keys2idl", "=> type %s indextype %s\n", type, indextype);
-    for (size_t i = 0; ivals[i] != NULL; i++) {
+    for (uint32_t i = 0; ivals[i] != NULL; i++) {
         IDList *idl2 = NULL;
 
         idl2 = index_read_ext_allids(pb, be, type, indextype, slapi_value_get_berval(ivals[i]), txn, err, unindexed, allidslimit);
@@ -985,8 +985,8 @@ keys2idl(
             char buf[BUFSIZ];
 
             slapi_log_err(SLAPI_LOG_TRACE, "keys2idl",
-                          "   ival[%lu] = \"%s\" => %lu IDs\n", i,
-                          encode(slapi_value_get_berval(ivals[i]), buf), (u_long)IDL_NIDS(idl2));
+                          "   ival[%" PRIu32 "] = \"%s\" => %" PRIu32 " IDs\n", i,
+                          encode(slapi_value_get_berval(ivals[i]), buf), (uint32_t)IDL_NIDS(idl2));
         }
 #endif
         if (idl2 == NULL) {
